Fix type error which GHC 7.8.3 missed
This program now compiles on GHC 7.10.1, but due to another bug, it is
effectively useless! This program:

  {-# LANGUAGE DataKinds #-}
  {-# LANGUAGE ConstraintKinds #-}
  {-# LANGUAGE TypeFamilies #-}

  import Data.TypeNat.Nat
  import Data.Proxy

  problem :: (StrongLTE (S Z) (S Z)) => Proxy m -> ()
  problem = const ()

  main = print (problem Proxy)

compiles just fine, ONLY IF you run GHCi from the TypeNat root
directory! Run it from elsewhere, and the type error is this:

  [1 of 1] Compiling Main             ( Ex.hs, Ex.o )

  Ex.hs:11:15:
      No instance for (LTE ('S 'Z) 'Z) arising from a use of ‘problem’
      In the first argument of ‘print’, namely ‘(problem Proxy)’
      In the expression: print (problem Proxy)
      In an equation for ‘main’: main = print (problem Proxy)

and this is incredibly weird. It looks like the LTE arguments were
flipped. I expect to see LTE 'Z ('S 'Z) in the constraint, but not the
other way around. Using GHCi's :kind! command, I have verified that the
constraint comes out as I expect it.

Some more diagnostics: if I copy the Nat.hs file to the same directory
as the problematic program above, and change the import and module names
suitably so that this file is used rather than that of TypeNat, then the
problem goes away. It seems that importing TypeNat from anywhere other
than the current working directory changes the beahviour of the type
checker.